do not lock shards while creating the temporary shard
